+++ src/textarea/Textarea.tsx
@@ -1,37 +1,67 @@
 import React, { forwardRef, useState, useEffect, useMemo, useRef, useCallback, useImperativeHandle } from 'react';
 import classNames from 'classnames';
-import useConfig from '../_util/useConfig';
+import parseTNode from '../_util/parseTNode';
 import useDefault from '../_util/useDefault';
-import { getCharacterLength } from '../_util/helper';
+import { getCharacterLength, limitUnicodeMaxLength } from '../_util/helper';
 import calcTextareaHeight from '../_common/js/utils/calcTextareaHeight';
+import { textareaDefaultProps } from './defaultProps';
 import { TdTextareaProps } from './type';
 import { StyledProps } from '../common';
+import { usePrefixClass } from '../hooks/useClass';
+import useDefaultProps from '../hooks/useDefaultProps';
 
-export interface TextareaProps extends TdTextareaProps, StyledProps {}
+export interface TextareaProps
+  extends Omit<
+      React.TextareaHTMLAttributes<HTMLTextAreaElement>,
+      'value' | 'defaultValue' | 'onBlur' | 'onChange' | 'onFocus'
+    >,
+    TdTextareaProps,
+    StyledProps {}
+
 export interface TextareaRefInterface extends React.RefObject<unknown> {
   currentElement: HTMLDivElement;
   textareaElement: HTMLTextAreaElement;
 }
 
-const Textarea = forwardRef((props: TextareaProps, ref: TextareaRefInterface) => {
-  const { disabled, maxlength, maxcharacter, autofocus, defaultValue, autosize = false, label, ...otherProps } = props;
-  const { classPrefix } = useConfig();
-  const baseClass = `${classPrefix}-textarea`;
+const Textarea = forwardRef<TextareaProps, TextareaRefInterface>((originProps, ref) => {
+  const props = useDefaultProps<TextareaProps>(originProps, textareaDefaultProps);
+  const {
+    className,
+    style,
+    allowInputOverMax,
+    autofocus,
+    bordered,
+    disabled,
+    defaultValue,
+    maxlength,
+    maxcharacter,
+    layout,
+    autosize,
+    label,
+    indicator,
+    readonly,
+    ...otherProps
+  } = props;
 
-  const [value = '', setValue] = useDefault(props.value, defaultValue, props.onChange);
+  const textareaClass = usePrefixClass('textarea');
+
+  const [value, setValue] = useDefault(props.value, defaultValue, props.onChange);
   const [textareaStyle, setTextareaStyle] = useState({});
+  const [composingValue, setComposingValue] = useState<string>('');
+  const composingRef = useRef(false);
   const textareaRef: React.RefObject<HTMLTextAreaElement> = useRef();
   const wrapperRef: React.RefObject<HTMLDivElement> = useRef();
 
   const textareaLength = useMemo(() => {
+    const realValue = composingRef.current ? composingValue : (value ?? '');
     if (typeof maxcharacter !== 'undefined') {
-      const { length = 0 } = getCharacterLength(String(value), maxcharacter) as {
+      const { length = 0 } = getCharacterLength(String(realValue), maxcharacter) as {
         length: number;
       };
       return length;
     }
-    return String(value).length || 0;
-  }, [value, maxcharacter]);
+    return String(realValue).length || 0;
+  }, [value, maxcharacter, composingRef, composingValue]);
 
   const textareaPropsNames = Object.keys(otherProps).filter((key) => !/^on[A-Z]/.test(key));
   const textareaProps = textareaPropsNames.reduce(
@@ -50,30 +80,65 @@ const Textarea = forwardRef((props: TextareaProps, ref: TextareaRefInterface) =>
     return eventProps;
   }, {});
 
-  const textareaClassNames = classNames(`${baseClass}__wrapper`, {
-    [`${baseClass}-is-disabled`]: disabled,
+  const textareaClasses = classNames(
+    `${textareaClass}`,
+    {
+      [`${textareaClass}--layout-${layout}`]: layout,
+      [`${textareaClass}--border`]: bordered,
+    },
+    className,
+  );
+
+  const textareaInnerClasses = classNames(`${textareaClass}__wrapper-inner`, {
+    [`${textareaClass}--disabled`]: disabled,
+    [`${textareaClass}--readonly`]: readonly,
   });
 
   const adjustTextareaHeight = useCallback(() => {
     if (autosize === true) {
       setTextareaStyle(calcTextareaHeight(textareaRef.current as HTMLTextAreaElement));
+    } else if (autosize === false) {
+      props.rows
+        ? setTextareaStyle({ height: 'auto', minHeight: 'auto' })
+        : setTextareaStyle(calcTextareaHeight(textareaRef.current as HTMLTextAreaElement, 1, 1));
     } else if (typeof autosize === 'object') {
       const { minRows, maxRows } = autosize;
       setTextareaStyle(calcTextareaHeight(textareaRef.current as HTMLTextAreaElement, minRows, maxRows));
-    } else if (autosize === false) {
-      setTextareaStyle({ height: 'auto', minHeight: '96px' });
     }
-  }, [autosize]);
+  }, [autosize, props.rows]);
 
-  function inputValueChangeHandle(e: React.FormEvent<HTMLTextAreaElement>) {
-    const { target } = e;
-    let val = (target as HTMLInputElement).value;
-    if (maxcharacter && maxcharacter >= 0) {
-      const stringInfo = getCharacterLength(val, maxcharacter);
-      val = typeof stringInfo === 'object' && stringInfo.characters;
+  const inputValueChangeHandle = (e: React.FormEvent<HTMLTextAreaElement>) => {
+    let { value: newStr } = e.target as HTMLInputElement;
+
+    if (value === newStr) return; // 避免在Firefox中重复触发
+
+    if (composingRef.current) {
+      setComposingValue(newStr);
+    } else {
+      if (!allowInputOverMax) {
+        newStr = limitUnicodeMaxLength(newStr, maxlength);
+        if (maxcharacter && maxcharacter >= 0) {
+          const stringInfo = getCharacterLength(newStr, maxcharacter);
+          newStr = typeof stringInfo === 'object' && stringInfo.characters;
+        }
+      }
+
+      // 中文输入结束，同步 composingValue
+      setComposingValue(newStr);
+      setValue(newStr, { e });
     }
-    setValue(val, { e });
-  }
+  };
+
+  const handleCompositionStart = () => {
+    composingRef.current = true;
+  };
+
+  const handleCompositionEnd = (e) => {
+    if (composingRef.current) {
+      composingRef.current = false;
+      inputValueChangeHandle(e);
+    }
+  };
 
   useEffect(() => {
     adjustTextareaHeight();
@@ -84,27 +149,40 @@ const Textarea = forwardRef((props: TextareaProps, ref: TextareaRefInterface) =>
     textareaElement: textareaRef.current,
   }));
 
+  const renderLabel = () => <div className={`${textareaClass}__label`}> {parseTNode(label)} </div>;
+
+  const renderIndicator = () => {
+    const isShowIndicator = indicator && (maxcharacter || maxlength);
+    if (!isShowIndicator) {
+      return null;
+    }
+    return <div className={`${textareaClass}__indicator`}>{`${textareaLength}/${maxcharacter || maxlength}`}</div>;
+  };
+
   return (
-    <div ref={wrapperRef} className={baseClass}>
-      {label && <div className={`${baseClass}__name`}> {label} </div>}
-      <div className={textareaClassNames}>
+    <div ref={wrapperRef} className={textareaClasses} style={style}>
+      {label && renderLabel()}
+      <div className={`${textareaClass}__wrapper`}>
         <textarea
           {...textareaProps}
           {...eventProps}
-          value={value}
+          className={textareaInnerClasses}
           style={textareaStyle}
+          value={composingRef.current ? composingValue : value}
+          readOnly={readonly}
           autoFocus={autofocus}
           disabled={disabled}
-          maxLength={maxlength}
           onChange={inputValueChangeHandle}
+          onCompositionStart={handleCompositionStart}
+          onCompositionEnd={handleCompositionEnd}
           ref={textareaRef}
         />
-        {(maxcharacter || maxlength) && (
-          <div className={`${baseClass}__count`}> {`${textareaLength}/${maxcharacter || maxlength}`} </div>
-        )}
+        {renderIndicator()}
       </div>
     </div>
   );
 });
+
+Textarea.displayName = 'Textarea';
 
 export default Textarea;