+++ src/textarea/_example/autosize.tsx
@@ -2,5 +2,5 @@ import React from 'react';
 import { Textarea } from 'tdesign-mobile-react';
 
 export default function Autosize() {
-  return <Textarea autosize placeholder="请输入文字" />;
+  return <Textarea placeholder="请输入文字" autosize />;
 }